+++ NEWS.md
@@ -1,6 +1,8 @@
 # ggh4x (development version)
 
-* `help_secondary()` has a new `na.rm` argument (#146, thanks @jrmuirhead!)
+* `help_secondary()` has a new `na.rm` argument (#146, thanks @jrmuirhead!).
+* Fixed a bug when using facets with extension elements (#151).
+* Added `at_panel()` for annotating specific panels in facets (#162)
 
 # ggh4x 0.2.8